requirements.txt
fastapi
uvicorn
pdfplumber
python-docx